+++ src/session.rs
@@ -1,9 +1,9 @@
-use crate::{AxumDatabasePool, AxumSessionData, AxumSessionID, AxumSessionStore, CookiesExt};
+use crate::{AxumDatabasePool, AxumSessionID, AxumSessionStore, CookiesExt};
 use async_trait::async_trait;
-use axum_core::extract::{FromRequest, RequestParts};
+use axum_core::extract::FromRequestParts;
 use cookie::CookieJar;
-use http::{self, StatusCode};
-use serde::{de::DeserializeOwned, Serialize};
+use http::{self, request::Parts, StatusCode};
+use serde::Serialize;
 use std::{
     fmt::Debug,
     marker::{Send, Sync},
@@ -24,19 +24,19 @@ where
     pub(crate) id: AxumSessionID,
 }
 
-/// Adds FromRequest<B> for AxumSession
+/// Adds FromRequestParts<B> for AxumSession
 ///
-/// Returns the AxumSession from Axums request extensions.
+/// Returns the AxumSession from Axums request extensions state.
 #[async_trait]
-impl<B, T> FromRequest<B> for AxumSession<T>
+impl<T, S> FromRequestParts<S> for AxumSession<T>
 where
-    B: Send,
     T: AxumDatabasePool + Clone + Debug + Sync + Send + 'static,
+    S: Send + Sync,
 {
     type Rejection = (http::StatusCode, &'static str);
 
-    async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
-        req.extensions().get::<AxumSession<T>>().cloned().ok_or((
+    async fn from_request_parts(parts: &mut Parts, _state: &S) -> Result<Self, Self::Rejection> {
+        parts.extensions.get::<AxumSession<T>>().cloned().ok_or((
             StatusCode::INTERNAL_SERVER_ERROR,
             "Can't extract AxumSession. Is `AxumSessionLayer` enabled?",
         ))
@@ -47,81 +47,96 @@ impl<S> AxumSession<S>
 where
     S: AxumDatabasePool + Clone + Debug + Sync + Send + 'static,
 {
-    pub(crate) fn new(store: &AxumSessionStore<S>, cookies: &CookieJar) -> AxumSession<S> {
+    pub(crate) async fn new(store: &AxumSessionStore<S>, cookies: &CookieJar) -> Self {
         let value = cookies
             .get_cookie(&store.config.cookie_name, &store.config.key)
             .and_then(|c| Uuid::parse_str(c.value()).ok());
 
-        let uuid = match value {
-            Some(v) => v,
-            None => loop {
-                let token = Uuid::new_v4();
-
-                if !store.inner.contains_key(&token.to_string()) {
-                    break token;
-                }
-            },
+        let id = match value {
+            Some(v) => AxumSessionID(v),
+            None => Self::generate_uuid(store).await,
         };
 
-        AxumSession {
-            id: AxumSessionID(uuid),
+        Self {
+            id,
             store: store.clone(),
         }
     }
-    /// Runs a Closure upon the Current Sessions stored data to get or set session data.
-    ///
-    /// Provides an Option<T> that returns the requested data from the Sessions store.
+
+    pub(crate) async fn generate_uuid(store: &AxumSessionStore<S>) -> AxumSessionID {
+        loop {
+            let token = Uuid::new_v4();
+
+            if !store.inner.contains_key(&token.to_string()) {
+                //This fixes an already used but in database issue.
+                if let Some(client) = &store.client {
+                    // Unwrap should be safe to use as we would want it to crash if there was a major database error.
+                    // This would mean the database no longer is online or the table missing etc.
+                    if !client
+                        .exists(&token.to_string(), &store.config.table_name)
+                        .await
+                        .unwrap()
+                    {
+                        return AxumSessionID(token);
+                    }
+                } else {
+                    return AxumSessionID(token);
+                }
+            }
+        }
+    }
+
+    /// Sets the Session to renew its Session ID.
+    /// This Deletes Session data from the database
+    /// associated with the old key.
     ///
     /// # Examples
-    /// ```rust no_run
-    /// session.tap(|sess| {
-    ///   let string = sess.data.get(key)?;
-    ///   serde_json::from_str(string).ok()
-    /// }).await;
+    /// ```rust ignore
+    /// session.destroy();
     /// ```
     ///
     #[inline]
-    pub(crate) fn tap<T: DeserializeOwned>(
-        &self,
-        func: impl FnOnce(&mut AxumSessionData) -> Option<T>,
-    ) -> Option<T> {
+    pub fn renew(&self) {
         if let Some(mut instance) = self.store.inner.get_mut(&self.id.0.to_string()) {
-            func(&mut instance)
+            instance.renew = true;
+            instance.update = true;
         } else {
             tracing::warn!("Session data unexpectedly missing");
-            None
         }
     }
 
     /// Sets the Current Session to be Destroyed on the next run.
     ///
     /// # Examples
-    /// ```rust no_run
-    /// session.destroy().await;
+    /// ```rust ignore
+    /// session.destroy();
     /// ```
     ///
     #[inline]
-    pub async fn destroy(&self) {
-        self.tap(|sess| {
-            sess.destroy = true;
-            Some(1)
-        });
+    pub fn destroy(&self) {
+        if let Some(mut instance) = self.store.inner.get_mut(&self.id.0.to_string()) {
+            instance.destroy = true;
+            instance.update = true;
+        } else {
+            tracing::warn!("Session data unexpectedly missing");
+        }
     }
 
     /// Sets the Current Session to a long term expiration. Useful for Remember Me setups.
     ///
     /// # Examples
-    /// ```rust no_run
-    /// session.set_longterm(true).await;
+    /// ```rust ignore
+    /// session.set_longterm(true);
     /// ```
     ///
     #[inline]
-    pub async fn set_longterm(&self, longterm: bool) {
-        self.tap(|sess| {
-            sess.longterm = longterm;
-            sess.update = true;
-            Some(1)
-        });
+    pub fn set_longterm(&self, longterm: bool) {
+        if let Some(mut instance) = self.store.inner.get_mut(&self.id.0.to_string()) {
+            instance.longterm = longterm;
+            instance.update = true;
+        } else {
+            tracing::warn!("Session data unexpectedly missing");
+        }
     }
 
     /// Sets the Current Session to be storable.
@@ -130,17 +145,18 @@ where
     /// If this is set to false it will unload the stored session.
     ///
     /// # Examples
-    /// ```rust no_run
-    /// session.set_store(true).await;
+    /// ```rust ignore
+    /// session.set_store(true);
     /// ```
     ///
     #[inline]
-    pub async fn set_store(&self, storable: bool) {
-        self.tap(|sess| {
-            sess.storable = storable;
-            sess.update = true;
-            Some(1)
-        });
+    pub fn set_store(&self, storable: bool) {
+        if let Some(mut instance) = self.store.inner.get_mut(&self.id.0.to_string()) {
+            instance.storable = storable;
+            instance.update = true;
+        } else {
+            tracing::warn!("Session data unexpectedly missing");
+        }
     }
 
     /// Gets data from the Session's HashMap
@@ -149,70 +165,98 @@ where
     /// Returns None if Key does not exist or if serdes_json failed to deserialize.
     ///
     /// # Examples
-    /// ```rust no_run
-    /// let id = session.get("user-id").await.unwrap_or(0);
+    /// ```rust ignore
+    /// let id = session.get("user-id").unwrap_or(0);
     /// ```
     ///
     ///Used to get data stored within SessionDatas hashmap from a key value.
     ///
     #[inline]
-    pub async fn get<T: serde::de::DeserializeOwned>(&self, key: &str) -> Option<T> {
-        self.tap(|sess| {
-            let string = sess.data.get(key)?;
-            serde_json::from_str(string).ok()
-        })
+    pub fn get<T: serde::de::DeserializeOwned>(&self, key: &str) -> Option<T> {
+        if let Some(instance) = self.store.inner.get_mut(&self.id.0.to_string()) {
+            let string = instance.data.get(key)?;
+            serde_json::from_str(&string).ok()
+        } else {
+            tracing::warn!("Session data unexpectedly missing");
+            None
+        }
+    }
+
+    /// Removes a Key from the Current Session's HashMap returning it.
+    ///
+    /// Provides an Option<T> that returns the requested data from the Sessions store.
+    /// Returns None if Key does not exist or if serdes_json failed to deserialize.
+    ///
+    /// # Examples
+    /// ```rust ignore
+    /// let id = session.get_remove("user-id").unwrap_or(0);
+    /// ```
+    ///
+    /// Used to get data stored within SessionDatas hashmap from a key value.
+    ///
+    #[inline]
+    pub fn get_remove<T: serde::de::DeserializeOwned>(&self, key: &str) -> Option<T> {
+        if let Some(mut instance) = self.store.inner.get_mut(&self.id.0.to_string()) {
+            let string = instance.data.remove(key)?;
+            instance.update = true;
+            serde_json::from_str(&string).ok()
+        } else {
+            tracing::warn!("Session data unexpectedly missing");
+            None
+        }
     }
 
     /// Sets data to the Current Session's HashMap.
     ///
     /// # Examples
-    /// ```rust no_run
-    /// session.set("user-id", 1).await;
+    /// ```rust ignore
+    /// session.set("user-id", 1);
     /// ```
     ///
     #[inline]
-    pub async fn set(&self, key: &str, value: impl Serialize) {
+    pub fn set(&self, key: &str, value: impl Serialize) {
         let value = serde_json::to_string(&value).unwrap_or_else(|_| "".to_string());
 
-        self.tap(|sess| {
-            if sess.data.get(key) != Some(&value) {
-                sess.data.insert(key.to_string(), value);
-                sess.update = true;
-            }
-            Some(1)
-        });
+        if let Some(mut instance) = self.store.inner.get_mut(&self.id.0.to_string()) {
+            let _ = instance.data.insert(key.to_string(), value);
+            instance.update = true;
+        } else {
+            tracing::warn!("Session data unexpectedly missing");
+        }
     }
 
     /// Removes a Key from the Current Session's HashMap.
+    /// Does not process the String into a Type, Just removes it.
     ///
     /// # Examples
-    /// ```rust no_run
-    /// session.remove("user-id").await;
+    /// ```rust ignore
+    /// let _ = session.remove("user-id");
     /// ```
     ///
     #[inline]
-    pub async fn remove(&self, key: &str) {
-        self.tap(|sess| {
-            sess.update = true;
-            sess.data.remove(key)
-        });
+    pub fn remove(&self, key: &str) {
+        if let Some(mut instance) = self.store.inner.get_mut(&self.id.0.to_string()) {
+            let _ = instance.data.remove(key);
+            instance.update = true;
+        } else {
+            tracing::warn!("Session data unexpectedly missing");
+        }
     }
 
     /// Clears all data from the Current Session's HashMap.
     ///
     /// # Examples
-    /// ```rust no_run
-    /// session.clear_all().await;
+    /// ```rust ignore
+    /// session.clear();
     /// ```
     ///
     #[inline]
-    pub async fn clear_all(&self) {
+    pub fn clear(&self) {
         if let Some(mut instance) = self.store.inner.get_mut(&self.id.0.to_string()) {
             instance.data.clear();
-        }
-
-        if self.store.is_persistent() {
-            self.store.clear_store().await.unwrap();
+            instance.update = true;
+        } else {
+            tracing::warn!("Session data unexpectedly missing");
         }
     }
 
@@ -222,7 +266,7 @@ where
     /// If the Session is not persistant it will return a count within AxumSessionStore.
     ///
     /// # Examples
-    /// ```rust no_run
+    /// ```rust ignore
     /// let count = session.count().await;
     /// ```
     ///